Server - refactor to no longer be a global, allow serving over HTTPS
The server's now refactored to be contained within its own module. If
the server port is set to 443, it will serve data over HTTPS instead of
HTTP. It is also no longer a global variable, now instead being a
module that can be required by the modules that use it.